requirements.txt
astroid==2.11.2
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
dill==0.3.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
numpy==1.22.3
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
pandas==1.4.3
Pillow==9.0.1
platformdirs==2.5.1
pycodestyle==2.8.0
pylint==2.13.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
scipy==1.8.0
six==1.16.0
sketchpy==0.0.15
svg.path==5.1
svgpathtools==1.4.4
svgwrite==1.4.2
toml==0.10.2
tqdm==4.63.1
turtle==0.0.1
urllib3==1.26.9
wrapt==1.14.0
